Fix MOAI zip file reading with 64bit systems, as well as small (< 256b) ...

+++ src/zl-vfs/ZLVfsZipArchive.h
@@ -4,20 +4,22 @@
 #ifndef ZLZIPARCHIVE_H
 #define ZLZIPARCHIVE_H
 
+#include <zl-common/zl_types.h>
+
 //================================================================//
 // ZLVfsZipArchiveHeader
 //================================================================//
 class ZLVfsZipArchiveHeader {
 public:
 
-	unsigned long	mSignature;			// 4 End of central directory signature = 0x06054b50
-	unsigned short	mDiskNumber;		// 2 Number of this disk
-	unsigned short	mStartDisk;			// 2 Disk where central directory starts
-	unsigned short	mTotalDiskEntries;	// 2 Total number of entries on disk
-	unsigned short	mTotalEntries;		// 2 Total number of central in archive
-	unsigned long	mCDSize;			// 4 Size of central directory in bytes
-	unsigned long	mCDAddr;			// 4 Offset of start of central directory, relative to start of archive
-	unsigned short	mCommentLength;		// 2 ZIP file comment length
+	u32	mSignature;			// 4 End of central directory signature = 0x06054b50
+	u16	mDiskNumber;		// 2 Number of this disk
+	u16	mStartDisk;			// 2 Disk where central directory starts
+	u16	mTotalDiskEntries;	// 2 Total number of entries on disk
+	u16	mTotalEntries;		// 2 Total number of central in archive
+	u32	mCDSize;			// 4 Size of central directory in bytes
+	u32	mCDAddr;			// 4 Offset of start of central directory, relative to start of archive
+	u16	mCommentLength;		// 2 ZIP file comment length
 	
 	//----------------------------------------------------------------//
 	int		FindAndRead		( FILE* file );
@@ -29,23 +31,23 @@ public:
 class ZLVfsZipEntryHeader {
 public:
 
-	unsigned long	mSignature;				// 4 Central directory file header signature = 0x02014b50
-	unsigned short	mByVersion;				// 2 Version made by
-	unsigned short	mVersionNeeded;			// 2 Version needed to extract (minimum)
-	unsigned short	mFlag;					// 2 General purpose bit flag
-	unsigned short	mCompression;			// 2 Compression method
-	unsigned short	mLastModTime;			// 2 File last modification time
-	unsigned short	mLastModDate;			// 2 File last modification date
-	unsigned long	mCrc32;					// 4 CRC-32
-	unsigned long	mCompressedSize;		// 4 Compressed size
-	unsigned long	mUncompressedSize;		// 4 Uncompressed size
-	unsigned short	mNameLength;			// 2 File name length (n)
-	unsigned short	mExtraFieldLength;		// 2 Extra field length (m)
-	unsigned short	mCommentLength;			// 2 File comment length (k)
-	unsigned short	mDiskNumber;			// 2 Disk number where file starts
-	unsigned short	mInternalAttributes;	// 2 Internal file attributes
-	unsigned long	mExternalAttributes;	// 4 External file attributes
-	unsigned long	mFileHeaderAddr;		// 4 Relative offset of file header
+	u32	mSignature;				// 4 Central directory file header signature = 0x02014b50
+	u16	mByVersion;				// 2 Version made by
+	u16	mVersionNeeded;			// 2 Version needed to extract (minimum)
+	u16	mFlag;					// 2 General purpose bit flag
+	u16	mCompression;			// 2 Compression method
+	u16	mLastModTime;			// 2 File last modification time
+	u16	mLastModDate;			// 2 File last modification date
+	u32	mCrc32;					// 4 CRC-32
+	u32	mCompressedSize;		// 4 Compressed size
+	u32	mUncompressedSize;		// 4 Uncompressed size
+	u16	mNameLength;			// 2 File name length (n)
+	u16	mExtraFieldLength;		// 2 Extra field length (m)
+	u16	mCommentLength;			// 2 File comment length (k)
+	u16	mDiskNumber;			// 2 Disk number where file starts
+	u16	mInternalAttributes;	// 2 Internal file attributes
+	u32	mExternalAttributes;	// 4 External file attributes
+	u32	mFileHeaderAddr;		// 4 Relative offset of file header
 	
 	//----------------------------------------------------------------//
 	int		Read	( FILE* file );
@@ -57,17 +59,17 @@ public:
 class ZLVfsZipFileHeader {
 public:
 
-	unsigned long	mSignature;				// 4	Local file header signature = 0x04034b50 (read as a little-endian number)
-	unsigned short	mVersionNeeded;			// 2	Version needed to extract (minimum)
-	unsigned short	mFlag;					// 2	General purpose bit flag
-	unsigned short	mCompression;			// 2	Compression method
-	unsigned short	mLastModTime;			// 2	File last modification time
-	unsigned short	mLastModDate;			// 2	File last modification date
-	unsigned long	mCrc32;					// 4	CRC-32 (*not* to be trusted - Android)
-	unsigned long	mCompressedSize;		// 4	Compressed size (*not* to be trusted - Android)
-	unsigned long	mUncompressedSize;		// 4	Uncompressed size (*not* to be trusted - Android)
-	unsigned short	mNameLength;			// 2	File name length
-	unsigned short	mExtraFieldLength;		// 2	Extra field length
+	u32	mSignature;				// 4	Local file header signature = 0x04034b50 (read as a little-endian number)
+	u16	mVersionNeeded;			// 2	Version needed to extract (minimum)
+	u16	mFlag;					// 2	General purpose bit flag
+	u16	mCompression;			// 2	Compression method
+	u16	mLastModTime;			// 2	File last modification time
+	u16	mLastModDate;			// 2	File last modification date
+	u32	mCrc32;					// 4	CRC-32 (*not* to be trusted - Android)
+	u32	mCompressedSize;		// 4	Compressed size (*not* to be trusted - Android)
+	u32	mUncompressedSize;		// 4	Uncompressed size (*not* to be trusted - Android)
+	u16	mNameLength;			// 2	File name length
+	u16	mExtraFieldLength;		// 2	Extra field length
 
 	//----------------------------------------------------------------//
 	int		Read	( FILE* file );
@@ -80,11 +82,11 @@ class ZLVfsZipFileEntry {
 public:
 
 	std::string		mName;
-	unsigned long	mFileHeaderAddr;
-	unsigned long	mCrc32;
-	unsigned short	mCompression;
-	unsigned long	mCompressedSize;
-	unsigned long	mUncompressedSize;
+	u32	mFileHeaderAddr;
+	u32	mCrc32;
+	u16	mCompression;
+	u32	mCompressedSize;
+	u32	mUncompressedSize;
 	
 	ZLVfsZipFileEntry*	mNext;
 };
